drivers: base: Make devcon functions return an ERR_PTR instead of NULL
When __device_find_connection or device_find_connection fails the caller
may want to wait if this is because there was no connection matching the
passed in dev + con-id in the connection table, or there was one but
looking up the data for the other end of the connection failed, which
typically means that the other device is not yet enumerated / probed.

This is esp. important so that drivers using devcon can properly return
-EPROBE_DEFER to wait for other drivers to load first.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>