usb: typec: mux: Return -EPROBE_DEFER on error from match functions
When getting a mux or switch and our conversion function gets called,
which means that there is a matching entry in the devcon connections
table, and we cannot find the mux or switch, return -EPROBE_DEFER to
let the caller know that the device should show up later.

This relies on the match in the devcon connections tabel being unique,
a single Type-C Port Controller will typically want to lookup both an
typec-mux and a typec-switch, specify a connection-id so that we have
unique matching entries in the devcon connection table for each.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>