usb: roles: Add Intel Cherrytrail USB role switch driver
Intel Cherrytrail SoCs have an internal USB role switch for swiching the
OTG USB data lines between the xHCI host controller and the dwc3 gadget
controller.

Note on some Cherrytrail systems the sw_id_pin bit of cfg0 is controlled
through AML code reacting on a GPIO IRQ connected to the USB OTG id pin
(through an _AIE ACPI method), but this driver is still needed to properly
set/clear vbus_valid so that gadget mode will work. On other systems this
driver is needed to control both the sw_id_pin cfg0 bit (which controls the
switch) and the vbus_valid bit.

The presence of AML code which modifies the cfg0 reg (on some systems)
means that we our read/write/modify of cfg0 may race with the AML code
doing the same to avoid this we take the global ACPI lock while doing
the read/write/modify.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>