usb: roles: Add usb_role_switch_disallow_userspace_control()
By default we allow userspace to control the role-switch for manual role
switching on devices with an USB micro-AB connector. But for Type-C ports
we do not want userspace to be able to control the role-switch.

Add an usb_role_switch_disallow_userspace_control() function which the
Type-C code can call to disallow userspace-control.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>