extcon: intel-int3496: Add support for controlling the CHT USB-role switch
The intel-int3496 extcon driver was original written for Bay Trail SoCs,
where an external mux, controlled through a GPIO is used to switch between
the USB host and device controllers.

Cherry Trail SoCs still use the INT3496 ACPI device to tell the OS about
the USB micro AB connector id-pin, but they have a built-in USB-role switch
for actually switching the USB data signals between the 2 controllers.

This commit adds support for using the new usb role-switch framework to
get a reference to this switch and switch it according to the id-pin value.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>