usb: roles: Allow role-switch driver to provide a get method
On some devices the role may be modified by AML code in that case we
want the sysfs attribute to reflect the actual role the hw is set to
not the cached value from the role-sw class device.

This commits add a get() method which role-sw drivers which can read
back the actual state from the hw can optionally provide.

This commit also removes the initial_role argument to
usb_role_switch_register() which I introduced earlier. Instead drivers
who can read-back the state should provide a get method.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>