usb: typec: wcove: Handle data role with the new USB Role API
So far the driver for WhiskeyCove PMIC USB Type-C PHY has
expected that firmware supplies ACPI _DSM method with a
function that configures the internal (Intel) USB role mux
accordingly. Unfortunately the ASL code for the function has
bugs. The biggest problem is that the xHCI PCI device is
not woken from D3 before the registers are accessed (the mux
is part of the Intel xHCI IP).

This will remove any dependency on firmware implemented ASL
code by relying on the new Intel USB Role mux driver, and
controlling the mux using the USB Role Class API instead of
calling the ACPI _DSM function.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>