extcon: axp288: Set USB role based on vbus-status where necessary
Ideally all devices with an AXP288 PMIC with a micro-AB USB OTG connector
would have an INT3496 ACPI device giving Linux access to the id-pin of
the connector and the USB role would be fully controlled by extcon-int3496
driver.

Unfortunately many devices do not have the INT3496 ACPI devices, instead
controlling the USB role through AML code, but the AML code typically only
switches between the host and none roles, because of Windows not really
using device mode.

To make device mode work this commit checks what mode the AML code has
set (so what value the id-pin has) and if it is not host mode switch
between the device / none roles based on the vbus-status.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>