usb: roles: Add usb_role_switch_get_role function
On some devices the usb_role_switch's role may be changed by the firmware,
in this case it is useful for usb_role_switch consumers to be able to read
the get the current role.

This commit adds an usb_role_switch_get_role function, which consumers can
use for this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>