extcon: axp288: Only reschedule charger-detection at boot when a SDP is detected
The only misdetection which can happen at boot due to data-lines mux issues
is detecting a non SDP as SDP, so we only need to retry if we detect a SDP
on our first detection.

Note Vbus misdetection is not a problem, as soon as the drivers controlling
the Vbus path set it correctly we will get an interrupt which reschedules
the charger-detection.

Also update the comment about the re-detection to reflect this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>