efi/x86: merge 32-bit and 64-bit UGA draw protocol setup routines
The two versions of setup_uga##() are mostly identical, with the
exception of the size of EFI_HANDLE. So let's merge the two, and
pull the implementation into the calling function setup_uga().

Note that the 64-bit version was only mixed-mode safe by accident:
it only calls the get_mode() method of the UGA draw protocol, which
happens to be the first member, and so truncating that value at
offset 0 to 32 bits happens to produce the correct value. But let's
not rely on that, and use the proper API instead.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>